requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django
django-filter==24.3
djangorestframework==3.15.2
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
httpcore==1.0.5
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
packaging==24.1
peft==0.12.0
pexpect==4.9.0
platformdirs==4.2.2
pox==0.3.4
ppft==1.7.6.8
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pybind11==2.13.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
ray==2.34.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
schema==0.7.7
scipy==1.14.0
sentry-sdk==2.13.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.2
tblib==3.0.0
tqdm==4.66.5
urllib3==2.2.2
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.0